@@ -33,13 +33,13 @@ class RecipientsDataTransformer implements DataTransformerInterface
      *
      * @return string
      */
-    public function transform($recipients)
+    public function transform($recipients): mixed
     {
         if (null === $recipients || 0 === $recipients->count()) {
             return '';
         }
 
-        $usernames = array();
+        $usernames = [];
 
         foreach ($recipients as $recipient) {
             $usernames[] = $this->userToUsernameTransformer->transform($recipient);
