@@ -7,17 +7,17 @@ use Symfony\Component\Security\Core\User\UserProviderInterface;
 
 class UserProvider implements UserProviderInterface
 {
-    public function loadUserByUsername($username)
+    public function loadUserByIdentifier($username): UserInterface
     {
         return new User();
     }
 
-    public function refreshUser(UserInterface $user)
+    public function refreshUser(UserInterface $user): UserInterface
     {
         return $user;
     }
 
-    public function supportsClass($class)
+    public function supportsClass($class): bool
     {
         return User::class === $class;
     }
@@ -25,5 +25,10 @@ class UserProvider implements UserProviderInterface
     private function fetchUser($username)
     {
         return new User();
+    }
+
+    public function loadUserByUsername($username): UserInterface
+    {
+        return $this->loadUserByIdentifier($username);
     }
 }
