@@ -16,7 +16,7 @@ class ThreadTest extends TestCase
         $thread = $this->getMockForAbstractClass('FOS\MessageBundle\Model\Thread');
         $thread->expects($this->atLeastOnce())
             ->method('getParticipants')
-            ->will($this->returnValue(array($u1, $u2, $u3)));
+            ->will($this->returnValue([$u1, $u2, $u3]));
 
         $toIds = function (array $participants) {
             return array_map(function (ParticipantInterface $participant) {
@@ -24,9 +24,9 @@ class ThreadTest extends TestCase
             }, $participants);
         };
 
-        $this->assertSame($toIds(array($u2, $u3)), $toIds($thread->getOtherParticipants($u1)));
-        $this->assertSame($toIds(array($u1, $u3)), $toIds($thread->getOtherParticipants($u2)));
-        $this->assertSame($toIds(array($u1, $u2)), $toIds($thread->getOtherParticipants($u3)));
+        $this->assertSame($toIds([$u2, $u3]), $toIds($thread->getOtherParticipants($u1)));
+        $this->assertSame($toIds([$u1, $u3]), $toIds($thread->getOtherParticipants($u2)));
+        $this->assertSame($toIds([$u1, $u2]), $toIds($thread->getOtherParticipants($u3)));
     }
 
     protected function createParticipantMock($id)
